+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 ### Added
 
- - Implement `HLPerformSelector` as an encodable alternative to
-   `[SKAction runBlock:]`.
-   [Karl Voskuil](https://github.com/karlvoskuil)
+- Implement in `HLAction.h` a number of encodable alternatives to
+  block-related `SKAction`s like `runBlock:`.  These allow running
+  custom animations through application state preservation and
+  restoration.
+  [Karl Voskuil](https://github.com/karlvoskuil)